To get the plot
This code is suppose to get values for the period in a certain range and the same number of values for binary mass in a given range. I took the stellar mass we have been working with and gave it a range around the actual value. Then the code should check which values will give deltaV values in the expected region (around the actual calculated value). It is giving 0 tolerance values. I assume it is either the model or the priors that are faulty. I tried different values and I tried to change the model with no luck.